// src/frontend/src/app/api/fn/test-cases/api-exams-exam-id-testcases-id-patch-json.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { TestCaseDto } from '../../models/test-case-dto';

export interface ApiExamsExamIdTestcasesIdPatch$Json$Params {
  examId: number;
  id: number;
      body?: {
'FilePath'?: string;
'File'?: Blob;
}
}

export function apiExamsExamIdTestcasesIdPatch$Json(http: HttpClient, rootUrl: string, params: ApiExamsExamIdTestcasesIdPatch$Json$Params, context?: HttpContext): Observable<StrictHttpResponse<TestCaseDto>> {
  const rb = new RequestBuilder(rootUrl, apiExamsExamIdTestcasesIdPatch$Json.PATH, 'patch');
  if (params) {
    rb.path('examId', params.examId, {});
    rb.path('id', params.id, {});
    rb.body(params.body, 'multipart/form-data');
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'text/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<TestCaseDto>;
    })
  );
}

apiExamsExamIdTestcasesIdPatch$Json.PATH = '/api/exams/{examId}/testcases/{id}';